// mobile-app/interfaces/weather.ts
import { CityData } from "./city";

export interface WeatherData {
  coordinate: Coordinate;
  description: string;
  temperature?: number;
  feelsLike?: number;
  humidity?: number;
  temperatureMin?: number;
  temperatureMax?: number;
  windSpeed?: number;
  timestamp?: number;
  timezone?: number;
  near: string;
  city?: CityData;
}

export interface Coordinate {
  longitude: number;
  latitude: number;
}

export interface CrudeWeatherData {
  coord: Coord;
  weather: Weather[];
  base: string;
  main: Main;
  visibility: number;
  wind: Wind;
  clouds: Clouds;
  dt: number;
  sys: Sys;
  timezone: number;
  id: number;
  name: string;
  cod: number;
}

interface Coord {
  lon: number;
  lat: number;
}

interface Weather {
  id: number;
  main: string;
  description: string;
  icon: string;
}

interface Main {
  temp: number;
  feels_like: number;
  temp_min: number;
  temp_max: number;
  pressure: number;
  humidity: number;
  sea_level: number;
  grnd_level: number;
}

interface Wind {
  speed: number;
  deg: number;
  gust: number;
}

interface Clouds {
  all: number;
}

interface Sys {
  sunrise: number;
  sunset: number;
}